+++ composer.json
@@ -32,9 +32,6 @@
   "require-dev": {
     "phpunit/phpunit": "^5.7"
   },
-  "require-dev": {
-    "phpunit/phpunit": "^5.7"
-  },
   "autoload": {
     "psr-4": {
       "Ichaber\\SSSwiftype\\": "src/",
