+++ tests/php/Extensions/SwiftypeMetaTagContentExtensionTest.php
@@ -11,6 +11,7 @@ use Ichaber\SSSwiftype\Tests\Fake\SwiftypeSiteTree;
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Dev\SapphireTest;
 use SilverStripe\ORM\FieldType\DBDatetime;
+use SilverStripe\ORM\FieldType\DBHTMLText;
 
 /**
  * Class SwiftypeMetaTagContentExtensionTest
@@ -61,5 +62,33 @@ class SwiftypeMetaTagContentExtensionTest extends SapphireTest
         $output = trim(preg_replace("/\s+/S", '', $page->getSwiftypeMetaTags()->getValue()));
 
         $this->assertEquals($mock, $output);
+    }
+
+    /**
+     * @throws Exception
+     */
+    public function testMetaTagsReturnsEmptyByDefault()
+    {
+        $page = new SwiftypeSiteTree();
+        /** @var DBHTMLText $tags */
+        $tags = $page->getSwiftypeMetaTags();
+        $this->assertEmpty($tags->getValue());
+    }
+
+    /**
+     * @throws Exception
+     */
+    public function testMetaTagsReturnsConfiguredTags()
+    {
+        $page = new SwiftypeSiteTree();
+        $page->MetaDescription = 'My description';
+        $page->config()->update('swiftype_meta_tag_classes', [
+            SwiftypeMetaTagDescription::class
+        ]);
+
+        /** @var DBHTMLText $tags */
+        $tags = $page->getSwiftypeMetaTags();
+        $this->assertNotEmpty($tags->getValue());
+        $this->assertContains($page->MetaDescription, $tags->getValue());
     }
 }
